requirements.txt
bcrypt==4.1.2
blinker==1.7.0
cachelib==0.12.0
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
defusedxml==0.7.1
Flask==3.0.2
Flask-Bcrypt==1.0.1
Flask-Session==0.6.0
fonttools==4.49.0
fpdf2==2.7.8
idna==3.6
itsdangerous==2.1.2
Jinja2==3.1.3
joblib==1.3.2
MarkupSafe==2.1.5
mysql-connector-python==8.3.0
networkx==3.2.1
nltk==3.8.1
numpy==1.26.4
pillow==10.2.0
regex==2023.12.25
requests==2.31.0
scikit-learn==1.4.1.post1
scipy==1.12.0
threadpoolctl==3.3.0
tqdm==4.66.2
urllib3==2.2.1
Werkzeug==3.0.1